+++ app/build.gradle
@@ -63,6 +63,15 @@ dependencies {
     androidTestImplementation 'com.android.support.test:rules:1.0.2'
     implementation 'com.android.support:design:28.0.0'
 
+    //mockito dependencies ------------------------------------------
+    testCompile 'junit:junit:4.12'
+    // required if you want to use Mockito for unit tests
+    testCompile 'org.mockito:mockito-core:2.7.22'
+    // required if you want to use Mockito for Android tests
+    androidTestCompile 'org.mockito:mockito-android:2.7.22'
+   // -----------------------------------------------------------------
+
+
     //-----------Required to use Room for Pre-AndroidX-------------
     //added by Jonathan when setting the project to use a local database
     def room_version = "1.1.1"
